Correcciones sobre buenas prácticas
